move test helper code into testutils, so it is not shipped with releases
