+++ tests/js/server/aql/aql-graph-traversal-generic-cluster.js
@@ -24,8 +24,8 @@
 /// @author Tobias Gödderz
 ////////////////////////////////////////////////////////////////////////////////
 
-const {protoGraphs} = require('@arangodb/aql-graph-traversal-generic-graphs.js');
-const {testsByGraph, metaTests} = require('@arangodb/aql-graph-traversal-generic-tests.js');
+const {protoGraphs} = require('@arangodb/testutils/aql-graph-traversal-generic-graphs.js');
+const {testsByGraph, metaTests} = require('@arangodb/testutils/aql-graph-traversal-generic-tests.js');
 
 const jsunity = require('jsunity');
 const console = require('console');
