+++ tests/js/server/aql/aql-graph-traversal-generic-noncluster.js
@@ -25,8 +25,8 @@
 /// @author Heiko Kernbach
 ////////////////////////////////////////////////////////////////////////////////
 
-const {protoGraphs} = require('@arangodb/aql-graph-traversal-generic-graphs.js');
-const {testsByGraph, metaTests} = require('@arangodb/aql-graph-traversal-generic-tests.js');
+const {protoGraphs} = require('@arangodb/testutils/aql-graph-traversal-generic-graphs.js');
+const {testsByGraph, metaTests} = require('@arangodb/testutils/aql-graph-traversal-generic-tests.js');
 
 const jsunity = require("jsunity");
 const _ = require("lodash");
